# example/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/{tenantID}/events": {
            "get": {
                "description": "Get all aggregate events (historical events and future patches) with optional search, sorting and pagination. The search is case-insensitive and has filter support via \"@ColumnName:X\". If no column is specified, all possible search fields are searched. Sorting for multiple columns via \"+ColumnName,-ColumnName2\" with +/- prefix to specify ascending/descending sort direction. The size parameter specifies the maximum number of entries that will be returned. With the values and direction parameters the start value of the page can be defined to achieve pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Get aggregate events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"2024-01-03T00:00:00Z\"",
                        "description": "Time (RFC3339)",
                        "name": "time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "@AggregateType:Item",
                        "description": "Search Term - supported columns are: EventID/SearchAggregateEventID, AggregateType/SearchAggregateType, AggregateID/SearchAggregateID, Version/SearchAggregateVersion, EventType/SearchAggregateEventType, Class/SearchAggregateClass, ValidTime/SearchValidTime, TransactionTime/SearchTransactionTime, Data/SearchData",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-ValidTime",
                        "description": "Sort Term - supported columns are: AggregateType/SortAggregateType, AggregateID/SortAggregateID, Version/SortAggregateVersion, EventType/SortAggregateEventType, Class/SortAggregateClass, ValidTime/SortValidTime, TransactionTime/SortTransactionTime",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1000,
                        "description": "Page Size - maximum 1000",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Values of first row on page (only for pagination)",
                        "name": "values",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "IsBackward (only for pagination)",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Found Events",
                        "schema": {
                            "$ref": "#/definitions/rest.EventStreamResponse"
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{tenantID}/events/{aggregateType}/{aggregateID}/{eventID}": {
            "delete": {
                "description": "Delete scheduled event of an aggregate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Delete scheduled event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Item\"",
                        "description": "Aggregate Type",
                        "name": "aggregateType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Aggregate ID",
                        "name": "aggregateID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "event.DTO": {
            "type": "object",
            "properties": {
                "aggregateID": {
                    "type": "string"
                },
                "aggregateType": {
                    "type": "string"
                },
                "editedBy": {
                    "type": "string"
                },
                "eventID": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "validFrom": {
                    "type": "string"
                },
                "valueChange": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "response.PageDTO": {
            "type": "object",
            "properties": {
                "backward": {
                    "type": "boolean"
                },
                "search": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "sort": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "rest.EventStreamResponse": {
            "type": "object",
            "properties": {
                "next": {
                    "$ref": "#/definitions/response.PageDTO"
                },
                "patches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/event.DTO"
                    }
                },
                "previous": {
                    "$ref": "#/definitions/response.PageDTO"
                },
                "validTime": {
                    "type": "string",
                    "example": "2024-01-03T00:00:00Z"
                }
            }
        },
...
    }
}